move file and add '?' at type Question
